requirements.txt
alembic==1.6.2
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==8.0.0
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
enums==0.0.2
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==3.1
itsdangerous==2.0.0
Jinja2==3.0.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.12.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
WTForms==2.3.3
